Grouped project version related tests
